+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2019-07-06
+
+### Added
+
+- builder now can generate nested set entities and nested set repositories
+
+### Changed
+
+- Kicked fos/user-bundle and implemented own user entity
+
 ## [0.1.0] - 2019-07-06
 
 ### Added
@@ -95,7 +105,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - base app
 
-[Unreleased]: https://github.com/siewert87/aaas-api/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/siewert87/aaas-api/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/siewert87/aaas-api/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/siewert87/aaas-api/compare/v0.0.9...v0.1.0
 [0.0.9]: https://github.com/siewert87/aaas-api/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/siewert87/aaas-api/compare/v0.0.7...v0.0.8